app/requirements.txt
Flask
psycopg2-binary
redis
flask-session
